requirements.txt
aiohttp[speedups]==3.8.4
anyio==3.7.1
orjson==3.9.2
SQLAlchemy==2.0.18
asyncpg==0.28.0
inflect==7.0.0
python-dateutil==2.8.2
asyncclick==8.1.3.4
aiogram==2.25.1